sentinel-apex-api/requirements.txt
# SENTINEL APEX API — Production Dependencies
# CYBERDUDEBIVASH PVT LTD

# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# HTTP Client (Supabase REST)
httpx==0.28.1

# Auth
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Security
cryptography==44.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
